feat(pwa): implement Progressive Web App
