+++ apps/jadwal-sholat/package.json
@@ -22,6 +22,7 @@
     "@apollo/client": "^3.8.8",
     "@apollo/experimental-nextjs-app-support": "^0.5.2",
     "@ducanh2912/next-pwa": "^9.7.2",
+    "@studio-freight/react-lenis": "^0.0.38",
     "@t3-oss/env-nextjs": "^0.7.1",
     "@tanstack/react-query": "^5.12.2",
     "axios": "^1.6.2",
@@ -33,7 +34,6 @@
     "next-themes": "^0.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-secure-storage": "^1.3.2",
     "react-string-replace": "^1.1.1",
     "react-use-audio-player": "^2.2.0",
     "tailwind-merge": "^2.1.0",
@@ -42,12 +42,18 @@
     "zustand": "^4.4.7"
   },
   "devDependencies": {
+    "@jadwal-sholat/eslint-config": "workspace:*",
+    "@jadwal-sholat/typescript-config": "workspace:*",
     "@types/node": "20.10.3",
     "@types/react": "18.2.42",
     "@types/react-dom": "18.2.17",
+    "@typescript-eslint/eslint-plugin": "^6.13.2",
+    "@typescript-eslint/parser": "^6.13.2",
+    "@unocss/postcss": "^0.58.0",
+    "@unocss/preset-wind": "^0.58.0",
+    "@unocss/reset": "^0.58.0",
     "cypress": "^13.6.0",
-    "eslint-config-custom": "workspace:*",
-    "tsconfig": "workspace:*",
-    "typescript": "5.3.2"
+    "typescript": "5.3.2",
+    "unocss": "^0.58.0"
   }
 }
