+++ apps/jadwal-sholat/public/manifest.json
@@ -2,8 +2,8 @@
   "theme_color": "#000",
   "background_color": "#ffffff",
   "display": "standalone",
-  "scope": "/",
   "start_url": "/",
+  "orientation": "portrait",
   "name": "Jadwal Sholat",
   "short_name": "jadwal-sholat",
   "description": "Aplikasi Web untuk menampilkan jadwal sholat, Asma'ul Husna, dan baca Al-Qur'an",
