+++ apps/jadwal-sholat/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import { Metadata } from "next";
+import { Metadata, Viewport } from "next";
 import { BackToTop } from "~components/atoms";
 import { Navbar } from "~components/organisms";
 import { ChildrenProps } from "~interfaces";
@@ -50,6 +50,12 @@ export const metadata: Metadata = {
     card: "summary_large_image",
   },
   metadataBase: new URL(url),
+};
+
+export const viewport: Viewport = {
+  width: "device-width",
+  initialScale: 1,
+  maximumScale: 1,
 };
 
 export default function Layout({ children }: ChildrenProps) {
