+++ apps/jadwal-sholat/src/app/wrapper.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { ReactLenis } from "@studio-freight/react-lenis";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ThemeProvider } from "next-themes";
 import { ChildrenProps } from "~interfaces";
@@ -10,7 +11,7 @@ export default function Wrapper({ children }: ChildrenProps) {
   return (
     <QueryClientProvider client={queryClient}>
       <ThemeProvider enableSystem attribute="class">
-        {children}
+        <ReactLenis root>{children}</ReactLenis>
       </ThemeProvider>
     </QueryClientProvider>
   );
