+++ apps/jadwal-sholat/src/components/organisms/detail-surat.tsx
@@ -3,7 +3,6 @@
 import { Pause, Play } from "lucide-react";
 import { nanoid } from "nanoid";
 import { memo } from "react";
-import secureLocalStorage from "react-secure-storage";
 import { ModalNotification, ModalTafsir } from "~components/molecules";
 import { usePlayNextAudio, useScrollAyat } from "~hooks";
 import { SuratProps } from "~interfaces";
@@ -37,7 +36,7 @@ export function DetailSurat({ surat }: SuratProps) {
   } = usePlayNextAudio(audioList);
 
   function saveData<T>(newData: T) {
-    secureLocalStorage.setItem("surat", JSON.stringify(newData));
+    localStorage.setItem("surat", JSON.stringify(newData));
   }
 
   function handleClick(name: string, ayat: number, number: number) {
