+++ apps/jadwal-sholat/src/components/organisms/index.ts
@@ -3,5 +3,4 @@ export { ListAsmaulHusna } from "./list-asmaul-husna";
 export { ListKota } from "./list-kota";
 export { ListSurat } from "./list-surat";
 export { Navbar } from "./navbar";
-export { TableJadwalPuasaSunnah } from "./table-jadwal-puasa-sunnah";
 export { TableJadwalSholat } from "./table-jadwal-sholat";
