+++ apps/jadwal-sholat/src/interfaces/asmaul-husna.interface.ts
@@ -1,0 +1,6 @@
+export interface AsmaulHusnaProps {
+  urutan: string;
+  latin: string;
+  arab: string;
+  arti: string;
+}
