+++ apps/jadwal-sholat/src/lib/helpers/remove-selected-surat.ts
@@ -1,8 +1,6 @@
-import secureLocalStorage from "react-secure-storage";
-
 // remove selected-surat item from localStorage
 export function removeSelectedSurat() {
-  if (secureLocalStorage.getItem("selected-surat") as string) {
-    secureLocalStorage.removeItem("selected-surat");
+  if (localStorage.getItem("selected-surat") as string) {
+    localStorage.removeItem("selected-surat");
   }
 }
