+++ packages/eslint-config/index.js
@@ -1,5 +1,11 @@
 module.exports = {
-  extends: ["next", "turbo", "next/core-web-vitals", "prettier", "plugin:jsx-a11y/recommended"],
+  extends: [
+    "next",
+    "turbo",
+    "next/core-web-vitals",
+    "prettier",
+    "plugin:jsx-a11y/recommended",
+  ],
   rules: {
     "@next/next/no-html-link-for-pages": "off",
   },
