+++ packages/eslint-config/package.json
@@ -1,19 +1,21 @@
 {
-  "name": "eslint-config-custom",
+  "name": "@jadwal-sholat/eslint-config",
   "version": "0.0.0",
-  "main": "index.js",
   "license": "MIT",
-  "dependencies": {
-    "eslint-config-turbo": "^1.9.3"
-  },
+  "files": [
+    "library.js",
+    "next.js",
+    "react-internal.js"
+  ],
   "publishConfig": {
     "access": "public"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
     "eslint-config-next": "^13.4.1",
     "eslint-config-prettier": "^8.8.0",
+    "eslint-config-turbo": "^1.9.3",
     "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-only-warn": "^1.1.0",
     "eslint-plugin-react": "7.28.0"
   }
 }
