+++ packages/typescript-config/base.json
