+++ packages/typescript-config/nextjs.json
