+++ packages/typescript-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tsconfig",
+  "name": "@jadwal-sholat/typescript-config",
   "version": "0.0.0",
   "private": true,
   "license": "MIT",
