+++ packages/typescript-config/react-library.json
