+++ prettier.config.js
@@ -4,7 +4,6 @@ module.exports = {
   endOfLine: "lf",
   htmlWhitespaceSensitivity: "css",
   insertPragma: false,
-  jsxBracketSameLine: false,
   jsxSingleQuote: false,
   printWidth: 80,
   proseWrap: "preserve",
