+++ tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@jadwal-sholat/typescript-config/base.json"
+}
